refactor: rename ga package to sdk
