+++ hrp/internal/sdk/client_test.go
@@ -1,4 +1,4 @@
-package ga
+package sdk
 
 import (
 	"testing"
