+++ hrp/internal/sdk/events.go
@@ -1,4 +1,4 @@
-package ga
+package sdk
 
 import (
 	"fmt"
